requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
faiss-cpu
pymupdf
langchain-google-genai
streamlit
python-dotenv
langchain-classic